# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Showroom Management Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface Product {
  id: number;
  name: string;
  sku: string;
  description?: string | null;
  price: number;
  cost: number;
  stock: number;
  minStock: number;
  categoryId?: number | null;
  categoryName?: string | null;
  brand?: string | null;
  imageUrl?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  sku: string;
  description?: string;
  price: number;
  cost: number;
  stock: number;
  minStock?: number;
  categoryId?: number;
  brand?: string;
  imageUrl?: string;
  isActive?: boolean;
}

export interface UpdateProductBody {
  name?: string;
  sku?: string;
  description?: string;
  price?: number;
  cost?: number;
  stock?: number;
  minStock?: number;
  categoryId?: number;
  brand?: string;
  imageUrl?: string;
  isActive?: boolean;
}

export interface ProductListResponse {
  items: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface Category {
  id: number;
  name: string;
  description?: string | null;
  productCount: number;
  createdAt: string;
}

export interface CreateCategoryBody {
  name: string;
  description?: string;
}

export interface Customer {
  id: number;
  name: string;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  totalOrders: number;
  totalSpent: number;
  createdAt: string;
}

export interface CreateCustomerBody {
  name: string;
  email?: string;
  phone?: string;
  address?: string;
}

export interface UpdateCustomerBody {
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
}

export interface CustomerListResponse {
  items: Customer[];
  total: number;
  page: number;
  limit: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  customerName: string;
  customerId?: number | null;
  status: OrderStatus;
  total: number;
  itemCount: number;
  notes?: string | null;
  createdAt: string;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  sku: string;
  quantity: number;
  unitPrice: number;
  subtotal: number;
}

export type OrderWithItems = Order & {
  items: OrderItem[];
};

export type CreateOrderBodyStatus =
  (typeof CreateOrderBodyStatus)[keyof typeof CreateOrderBodyStatus];

export const CreateOrderBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type CreateOrderBodyItemsItem = {
  productId: number;
  quantity: number;
  unitPrice: number;
};

export interface CreateOrderBody {
  customerId?: number;
  customerName: string;
  status?: CreateOrderBodyStatus;
  notes?: string;
  items: CreateOrderBodyItemsItem[];
}

export type UpdateOrderBodyStatus =
  (typeof UpdateOrderBodyStatus)[keyof typeof UpdateOrderBodyStatus];

export const UpdateOrderBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderBody {
  status?: UpdateOrderBodyStatus;
  notes?: string;
}

export interface OrderListResponse {
  items: Order[];
  total: number;
  page: number;
  limit: number;
}

export type StaffRole = (typeof StaffRole)[keyof typeof StaffRole];

export const StaffRole = {
  manager: "manager",
  sales: "sales",
  technician: "technician",
  support: "support",
} as const;

export interface Staff {
  id: number;
  name: string;
  email: string;
  role: StaffRole;
  phone?: string | null;
  isActive: boolean;
  createdAt: string;
}

export type CreateStaffBodyRole =
  (typeof CreateStaffBodyRole)[keyof typeof CreateStaffBodyRole];

export const CreateStaffBodyRole = {
  manager: "manager",
  sales: "sales",
  technician: "technician",
  support: "support",
} as const;

export interface CreateStaffBody {
  name: string;
  email: string;
  role: CreateStaffBodyRole;
  phone?: string;
}

export type UpdateStaffBodyRole =
  (typeof UpdateStaffBodyRole)[keyof typeof UpdateStaffBodyRole];

export const UpdateStaffBodyRole = {
  manager: "manager",
  sales: "sales",
  technician: "technician",
  support: "support",
} as const;

export interface UpdateStaffBody {
  name?: string;
  email?: string;
  role?: UpdateStaffBodyRole;
  phone?: string;
  isActive?: boolean;
}

export interface DashboardStats {
  totalRevenue: number;
  revenueGrowth: number;
  totalOrders: number;
  ordersGrowth: number;
  totalCustomers: number;
  customersGrowth: number;
  totalProducts: number;
  lowStockCount: number;
  pendingOrders: number;
  completedOrdersToday: number;
}

export interface RevenueChartPoint {
  month: string;
  revenue: number;
  orders: number;
}

export interface TopProduct {
  id: number;
  name: string;
  sku: string;
  totalSold: number;
  revenue: number;
  categoryName?: string | null;
}

export interface CategoryBreakdown {
  categoryName: string;
  revenue: number;
  orderCount: number;
  percentage: number;
}

export type ListProductsParams = {
  search?: string;
  categoryId?: number;
  lowStock?: boolean;
  page?: number;
  limit?: number;
};

export type ListCustomersParams = {
  search?: string;
  page?: number;
  limit?: number;
};

export type ListOrdersParams = {
  status?: ListOrdersStatus;
  customerId?: number;
  page?: number;
  limit?: number;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type GetRevenueChartParams = {
  months?: number;
};

export type GetRecentOrdersParams = {
  limit?: number;
};

export type GetTopProductsParams = {
  limit?: number;
};