# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Showroom Management Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  customerName: string;
  customerId?: number | null;
  status: OrderStatus;
  total: number;
  itemCount: number;
  notes?: string | null;
  createdAt: Date;
}